reset motor_step at 4 instead of each rotation
This has the same effect but requires less math to calculate the pin
config

+++ MultiStepper.h
@@ -62,7 +62,7 @@ class MultiStepper {
     volatile uint8_t *limit_port;
 
     //motor state data
-    int motor_step[4];
+    uint8_t motor_step[4];
     long motor_position[4];
 };
